Apply more refactoring
