--- lib/streamlit/elements/arrow_vega_lite.py
@@ -1,229 +1,0 @@
-# Copyright (c) Streamlit Inc. (2018-2022) Snowflake Inc. (2022-2024)
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""A Python wrapper around Vega-Lite."""
-
-from __future__ import annotations
-
-import json
-from typing import TYPE_CHECKING, Any, Literal, cast
-
-import streamlit.elements.lib.dicttools as dicttools
-from streamlit.elements import arrow
-from streamlit.elements.arrow import Data
-from streamlit.errors import StreamlitAPIException
-from streamlit.proto.ArrowVegaLiteChart_pb2 import (
-    ArrowVegaLiteChart as ArrowVegaLiteChartProto,
-)
-from streamlit.runtime.metrics_util import gather_metrics
-
-if TYPE_CHECKING:
-    from streamlit.delta_generator import DeltaGenerator
-
-
-class ArrowVegaLiteMixin:
-    @gather_metrics("vega_lite_chart")
-    def vega_lite_chart(
-        self,
-        data: Data = None,
-        spec: dict[str, Any] | None = None,
-        use_container_width: bool = False,
-        theme: Literal["streamlit"] | None = "streamlit",
-        **kwargs: Any,
-    ) -> DeltaGenerator:
-        """Display a chart using the Vega-Lite library.
-
-        Parameters
-        ----------
-        data : pandas.DataFrame, pandas.Styler, pyarrow.Table, numpy.ndarray, Iterable, dict, or None
-            Either the data to be plotted or a Vega-Lite spec containing the
-            data (which more closely follows the Vega-Lite API).
-
-        spec : dict or None
-            The Vega-Lite spec for the chart. If the spec was already passed in
-            the previous argument, this must be set to None. See
-            https://vega.github.io/vega-lite/docs/ for more info.
-
-        use_container_width : bool
-            If True, set the chart width to the column width. This takes
-            precedence over Vega-Lite's native `width` value.
-
-        theme : "streamlit" or None
-            The theme of the chart. Currently, we only support "streamlit" for the Streamlit
-            defined design or None to fallback to the default behavior of the library.
-
-        **kwargs : any
-            Same as spec, but as keywords.
-
-        Example
-        -------
-        >>> import streamlit as st
-        >>> import pandas as pd
-        >>> import numpy as np
-        >>>
-        >>> chart_data = pd.DataFrame(np.random.randn(200, 3), columns=["a", "b", "c"])
-        >>>
-        >>> st.vega_lite_chart(
-        ...    chart_data,
-        ...    {
-        ...        "mark": {"type": "circle", "tooltip": True},
-        ...        "encoding": {
-        ...            "x": {"field": "a", "type": "quantitative"},
-        ...            "y": {"field": "b", "type": "quantitative"},
-        ...            "size": {"field": "c", "type": "quantitative"},
-        ...            "color": {"field": "c", "type": "quantitative"},
-        ...        },
-        ...    },
-        ... )
-
-        .. output::
-           https://doc-vega-lite-chart.streamlit.app/
-           height: 300px
-
-        Examples of Vega-Lite usage without Streamlit can be found at
-        https://vega.github.io/vega-lite/examples/. Most of those can be easily
-        translated to the syntax shown above.
-
-        """
-        if theme != "streamlit" and theme != None:
-            raise StreamlitAPIException(
-                f'You set theme="{theme}" while Streamlit charts only support theme=”streamlit” or theme=None to fallback to the default library theme.'
-            )
-        proto = ArrowVegaLiteChartProto()
-        marshall(
-            proto,
-            data,
-            spec,
-            use_container_width=use_container_width,
-            theme=theme,
-            **kwargs,
-        )
-        return self.dg._enqueue("arrow_vega_lite_chart", proto)
-
-    @property
-    def dg(self) -> DeltaGenerator:
-        """Get our DeltaGenerator."""
-        return cast("DeltaGenerator", self)
-
-
-def marshall(
-    proto: ArrowVegaLiteChartProto,
-    data: Data = None,
-    spec: dict[str, Any] | None = None,
-    use_container_width: bool = False,
-    theme: None | Literal["streamlit"] = "streamlit",
-    **kwargs,
-):
-    """Construct a Vega-Lite chart object.
-
-    See DeltaGenerator.vega_lite_chart for docs.
-    """
-    # Support passing data inside spec['datasets'] and spec['data'].
-    # (The data gets pulled out of the spec dict later on.)
-    if isinstance(data, dict) and spec is None:
-        spec = data
-        data = None
-
-    # Support passing no spec arg, but filling it with kwargs.
-    # Example:
-    #   marshall(proto, baz='boz')
-    if spec is None:
-        spec = dict()
-    else:
-        # Clone the spec dict, since we may be mutating it.
-        spec = dict(spec)
-
-    # Support passing in kwargs. Example:
-    #   marshall(proto, {foo: 'bar'}, baz='boz')
-    if len(kwargs):
-        # Merge spec with unflattened kwargs, where kwargs take precedence.
-        # This only works for string keys, but kwarg keys are strings anyways.
-        spec = dict(spec, **dicttools.unflatten(kwargs, _CHANNELS))
-
-    if len(spec) == 0:
-        raise ValueError("Vega-Lite charts require a non-empty spec dict.")
-
-    if "autosize" not in spec:
-        # type fit does not work for many chart types. This change focuses
-        # on vconcat with use_container_width=True as there are unintended
-        # consequences of changing the default autosize for all charts.
-        # fit-x fits the width and height can be adjusted.
-        if "vconcat" in spec and use_container_width:
-            spec["autosize"] = {"type": "fit-x", "contains": "padding"}
-        else:
-            spec["autosize"] = {"type": "fit", "contains": "padding"}
-
-    # Pull data out of spec dict when it's in a 'datasets' key:
-    #   marshall(proto, {datasets: {foo: df1, bar: df2}, ...})
-    if "datasets" in spec:
-        for k, v in spec["datasets"].items():
-            dataset = proto.datasets.add()
-            dataset.name = str(k)
-            dataset.has_name = True
-            arrow.marshall(dataset.data, v)
-        del spec["datasets"]
-
-    # Pull data out of spec dict when it's in a top-level 'data' key:
-    #   marshall(proto, {data: df})
-    #   marshall(proto, {data: {values: df, ...}})
-    #   marshall(proto, {data: {url: 'url'}})
-    #   marshall(proto, {data: {name: 'foo'}})
-    if "data" in spec:
-        data_spec = spec["data"]
-
-        if isinstance(data_spec, dict):
-            if "values" in data_spec:
-                data = data_spec["values"]
-                del spec["data"]
-        else:
-            data = data_spec
-            del spec["data"]
-
-    proto.spec = json.dumps(spec)
-    proto.use_container_width = use_container_width
-    proto.theme = theme or ""
-
-    if data is not None:
-        arrow.marshall(proto.data, data)
-
-
-# See https://vega.github.io/vega-lite/docs/encoding.html
-_CHANNELS = {
-    "x",
-    "y",
-    "x2",
-    "y2",
-    "xError",
-    "xError2",
-    "yError",
-    "yError2",
-    "longitude",
-    "latitude",
-    "color",
-    "opacity",
-    "fillOpacity",
-    "strokeOpacity",
-    "strokeWidth",
-    "size",
-    "shape",
-    "text",
-    "tooltip",
-    "href",
-    "key",
-    "order",
-    "detail",
-    "facet",
-    "row",
-    "column",
-}
